src/test/java/com/ryno/tp/immutablewarehouseretailer/model/employeesTest/ManagerTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.ryno.tp.immutablewarehouseretailer.model.employeesTest;

import com.ryno.tp.immutablewarehouseretailer.model.employees.Manager;
import com.ryno.tp.immutablewarehouseretailer.model.product.Appliance;
import org.testng.Assert;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author ryno
 */
public class ManagerTest {
    
    public ManagerTest() {
    }
    // TODO add test methods here.
    // The methods must be annotated with annotation @Test. For example:
    //
    // @Test
    // public void hello() {}

    @Test
    public void testCreation(){
        
        Manager man = new Manager.Builder("Ryno","Man007", 1500).Build();
        Assert.assertEquals(man.getEmpNum(),"Man007");
        
    }
    
    @Test
    public void testUpdate() throws Exception{
        
        Manager manB = new Manager.Builder("Emily", "Man008", 2000).Build();
        Assert.assertEquals(manB.getEmpNum(),"Man008");
    }
    
    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}